Refactor to improve structure
