--- ai_launchpad/data_models.py
@@ -1,8 +1,0 @@
-from typing import TypedDict
-
-
-class WorkItem(TypedDict):
-    title: str
-    description: str
-    link: str
-    relevant_source_directories: list[str]
