+++ tests/test_config.py
@@ -1,5 +1,4 @@
 import pytest
-from pathlib import Path
 from unittest.mock import patch
 
 from ai_launchpad.config import _read_toml, _missing_required_fields, read_config
@@ -48,7 +47,9 @@ class TestReadConfig:
         config_file.write_text(
             'base_worktrees_dir = "/tmp/ctx"\nbase_source_dir = "/tmp/src"\n'
         )
-        with patch("ai_launchpad.config._default_config_path", return_value=config_file):
+        with patch(
+            "ai_launchpad.config._default_config_path", return_value=config_file
+        ):
             result = read_config()
         assert result["base_worktrees_dir"] == "/tmp/ctx"
         assert result["base_source_dir"] == "/tmp/src"
@@ -56,13 +57,17 @@ class TestReadConfig:
     def test_empty_config_raises(self, tmp_path):
         config_file = tmp_path / "config.toml"
         config_file.write_text("")
-        with patch("ai_launchpad.config._default_config_path", return_value=config_file):
+        with patch(
+            "ai_launchpad.config._default_config_path", return_value=config_file
+        ):
             with pytest.raises(ValueError, match="empty or invalid"):
                 read_config()
 
     def test_missing_required_fields_raises(self, tmp_path):
         config_file = tmp_path / "config.toml"
         config_file.write_text('base_worktrees_dir = "/tmp/ctx"\n')
-        with patch("ai_launchpad.config._default_config_path", return_value=config_file):
+        with patch(
+            "ai_launchpad.config._default_config_path", return_value=config_file
+        ):
             with pytest.raises(ValueError, match="base_source_dir"):
                 read_config()
