+++ tests/test_work_items.py
@@ -1,4 +1,4 @@
-from ai_launchpad.data_models import WorkItem
+from ai_launchpad.work_items import WorkItem
 
 
 class TestWorkItem:
